cleanup about page
